add appointments model
