+++ src/resources/users/user-model.js
@@ -75,22 +75,6 @@ async function get_coaches() {
   return coaches;
 }
 
-async function get_appointments(role, id) {
-  let appointments;
-  if (role === 1) {
-    appointments = await db('users')
-      .join('students', 'students.user_id', '=', 'users.id')
-      .join('appointments', 'appointments.user_id', '=', 'users.id')
-      .where('users.id', '=', id)
-  } else {
-    appointments = await db('users')
-      .join('coaches', 'coaches.user_id', '=', 'users.id')
-      .join('appointments', 'appointments.user_id', '=', 'users.id')
-      .where('users.id', '=', id)
-  }
-  return appointments;
-}
-
 module.exports = {
   find,
   findBy,
@@ -99,5 +83,4 @@ module.exports = {
   remove,
   user_details,
   get_coaches,
-  get_appointments
 };
